Added a travis config and test launcher script
The travis config installs mongo2.6 and adds the 'eve' user to the database
before running the tests.

A script run_server_and_tests.py runs the server (run.py) and then in parallel
launches the tests (tests.py), and shuts down the server when finished. The
travis config uses this to run the tests in the CI environment.

Also added a requirements.txt file

Finally, clarified a comment about access control